Rename core to common
