+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode/*
 !.vscode/c_cpp_properties.json
+!.vscode/launch.json
 build/
