+++ CMakeLists.txt
@@ -14,11 +14,6 @@ set(GLFW_BUILD_DOCS OFF CACHE BOOL "" FORCE)
 set(GLFW_BUILD_TESTS OFF CACHE BOOL "" FORCE)
 set(GLFW_BUILD_EXAMPLES OFF CACHE BOOL "" FORCE)
 
-# File structure of the repo is organized
-# in such a way that a single project can
-# include any header from any other project.
-set(GIT_REPO_ROOT "${CMAKE_CURRENT_LIST_DIR}")
-
 add_subdirectory(deps/bitsery)
 add_subdirectory(deps/enet)
 add_subdirectory(deps/entt)
@@ -31,6 +26,6 @@ add_subdirectory(deps/stb)
 add_subdirectory(deps/thread_pool)
 add_subdirectory(deps/toml)
 
-add_subdirectory(core)
 add_subdirectory(client)
+add_subdirectory(common)
 add_subdirectory(main)
