+++ client/config.hpp
@@ -5,8 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 #pragma once
-#include <core/config.hpp>
-#include <core/types.hpp>
+#include <common/config.hpp>
+#include <common/types.hpp>
 
 class ClientConfig final : public TomlConfig<ClientConfig> {
 public:
