+++ common/config.hpp
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 #pragma once
-#include <core/fs.hpp>
+#include <common/fs.hpp>
 #include <spdlog/spdlog.h>
 #include <stdexcept>
 #include <sstream>
