+++ common/math/CMakeLists.txt
@@ -1,3 +1,3 @@
-target_sources(core PRIVATE
+target_sources(common PRIVATE
     "${CMAKE_CURRENT_LIST_DIR}/crc64.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/frustum.cpp")
