+++ common/math/crc64.cpp
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#include <core/math/crc64.hpp>
+#include <common/math/crc64.hpp>
 
 // I understand that having the table pre-generated
 // is a potentional safety risk but CRC64 is not going
